Moved GrowthMonitoringFormsAggregationHelper to own file
