+++ custom/icds_reports/models/aggregate.py
@@ -16,7 +16,7 @@ from custom.icds_reports.utils.aggregation import (
     BirthPreparednessFormsAggregationHelper,
     CcsRecordMonthlyAggregationHelper, ChildHealthMonthlyAggregationHelper,
     DailyFeedingFormsChildHealthAggregationHelper,
-    DeliveryFormsAggregationHelper, GrowthMonitoringFormsAggregationHelper,
+    DeliveryFormsAggregationHelper,
     InactiveAwwsAggregationHelper,
     AwwIncentiveAggregationHelper
 )
@@ -27,6 +27,7 @@ from six.moves import range
 from custom.icds_reports.utils.aggregation_helpers.complementary_forms import ComplementaryFormsAggregationHelper
 from custom.icds_reports.utils.aggregation_helpers.complementary_forms_ccs_record import \
     ComplementaryFormsCcsRecordAggregationHelper
+from custom.icds_reports.utils.aggregation_helpers.growth_monitoring_forms import GrowthMonitoringFormsAggregationHelper
 from custom.icds_reports.utils.aggregation_helpers.postnatal_care_forms_ccs_record import \
     PostnatalCareFormsCcsRecordAggregationHelper
 from custom.icds_reports.utils.aggregation_helpers.postnatal_care_forms_child_health import \
